// frontend/src/api/generated/model/productSummaryResponseDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ProductSummaryResponseDtoStatus = typeof ProductSummaryResponseDtoStatus[keyof typeof ProductSummaryResponseDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductSummaryResponseDtoStatus = {
  UNREGISTERED: 'UNREGISTERED',
  READY: 'READY',
  IN_PROGRESS: 'IN_PROGRESS',
  ENDED: 'ENDED',
} as const;